+++ python/crime_rate.py
@@ -1,9 +1,10 @@
 import pandas as pd
 import os
-from abbreviate_state import *
-from pymongo import MongoClient
+from data_utils import *
+from pymongo import MongoClient, UpdateOne
 from dotenv import load_dotenv
 
+# Load environment variables
 load_dotenv("../.env")
 
 # Connect to DB
@@ -24,10 +25,13 @@ df['crime'] = (df['Violent\ncrime'].astype(float) / df['Population'].astype(floa
 df = df[['state', 'name', 'crime']]
 
 # Update database
+writes = []
 for city in cities.find():
-    name = city["name"].replace("St.", "Saint").replace("town", "")
+    name = city["name"].replace("town", "").replace(" City", "")
     row = df[df['name'].str.contains("(?<![A-z])" + name) & df['state'].str.contains(city['state'])]
     if not row.empty:
         query = {'name': city['name'], 'state': city['state']}
         values = { "$set": { "crime_rate": row.values[0][2]}}
-        cities.update_one(query, values)
+        writes.append(UpdateOne(query, values))
+
+cities.bulk_write(writes)
