+++ src/components/Header.tsx
@@ -93,12 +93,13 @@ const Header = () => {
         {
             term: "RPP",
             description: "Regional Price Parity (RPP) is a measure of the differences in the price levels of goods and services across states and metropolitan areas for a given year." +
-                " RPPs are expressed as a percentage of the overall national price level for each year."
+                " RPPs are expressed as a percentage of the overall national price level for each year. Due to smaller sample sizes, a portion of our cities use state RPPs."
         },
         {
             term: "Outdoor Score",
             description: "The outdoor score is calculated using a weighted, curved average using the percentage of the state covered by GAP-1, GAP-2, and GAP-3 protected areas, percentage of the state" +
-            " economy made up by the outdoor recreation industry, median annual air quality, LBs of chemicals released annually per state square mile, and miles of trail per state square mile."
+            " economy made up by the outdoor recreation industry, median annual air quality, LBs of chemicals released annually per state square mile, and miles of trail per state square mile." +
+            "Due to smaller sample sizes, a portion of our cities use outdoor scores without the median annual air quality (state score)."
         },
         {
             term: "Description",
