--- src/types/Ranges.ts
@@ -1,24 +1,0 @@
-export type Ranges = {
-"_id":null,
-"max_population": number,
-"min_population": number,
-"avg_population": number,
-"max_density": number,
-"min_density": number,
-"avg_density": number,
-"max_age": number,
-"min_age": number,
-"avg_age": number,
-"max_summer_temp": number,
-"min_summer_temp": number,
-"avg_summer_temp": number,
-"max_winter_temp": number,
-"min_winter_temp": number,
-"avg_winter_temp": number,
-"max_rain": number,
-"min_rain": number,
-"avg_rain": number,
-"max_snow": number,
-"min_snow": number,
-"avg_snow": number
-}
